refactor: header 컴포넌트 분리 및 UI 개선
