+++ src/widgets/header/ui/HeaderInner.tsx
@@ -1,0 +1,125 @@
+'use client';
+
+import { usePathname, useRouter, useSearchParams } from 'next/navigation';
+import { useTranslations } from 'next-intl';
+import { useLanguageStore } from '@/shared/model';
+import { useUserStore } from '@/entities/user';
+import { useState, useEffect, useRef } from 'react';
+import DesktopHeader from './DesktopHeader';
+import MobileHeader from './MobileHeader';
+import MobileTopBar from './MobileTopBar';
+
+export function HeaderInner() {
+  const pathname = usePathname();
+  const router = useRouter();
+  const searchParams = useSearchParams();
+  const t = useTranslations('nav');
+  const { language, setLanguage } = useLanguageStore();
+  const { logout } = useUserStore();
+  const [isLangOpen, setIsLangOpen] = useState(false);
+  const [isSearchOpen, setIsSearchOpen] = useState(false);
+  const mobileHeaderRef = useRef<HTMLElement>(null);
+
+  const [searchValue, setSearchValue] = useState(searchParams.get('q') ?? '');
+
+  // Close search on outside click
+  useEffect(() => {
+    function handleClickOutside(event: MouseEvent) {
+      const target = event.target as Node;
+      if (mobileHeaderRef.current && !mobileHeaderRef.current.contains(target)) {
+        setIsSearchOpen(false);
+      }
+    }
+    if (isSearchOpen) {
+      document.addEventListener('mousedown', handleClickOutside);
+    }
+    return () => document.removeEventListener('mousedown', handleClickOutside);
+  }, [isSearchOpen]);
+
+  // Close lang dropdown on Escape
+  useEffect(() => {
+    if (!isLangOpen) return;
+    function handleKeyDown(e: KeyboardEvent) {
+      if (e.key === 'Escape') setIsLangOpen(false);
+    }
+    document.addEventListener('keydown', handleKeyDown);
+    return () => document.removeEventListener('keydown', handleKeyDown);
+  }, [isLangOpen]);
+
+  const isHome = pathname === '/';
+  const isKo = language === 'ko';
+  const activeCategory = searchParams.get('category') ?? '전체';
+
+  function handleCategoryClick(cat: string) {
+    const params = new URLSearchParams(searchParams.toString());
+    if (cat === '전체') {
+      params.delete('category');
+    } else {
+      params.set('category', cat);
+    }
+    router.push(`/?${params.toString()}`, { scroll: false });
+  }
+
+  function handleSearch(e: React.FormEvent) {
+    e.preventDefault();
+    const params = new URLSearchParams(searchParams.toString());
+    if (searchValue.trim()) {
+      params.set('q', searchValue.trim());
+    } else {
+      params.delete('q');
+    }
+    router.push(`/?${params.toString()}`);
+    setIsSearchOpen(false);
+  }
+
+  function handleLogout() {
+    logout();
+    router.push('/');
+  }
+
+  const langs: { code: 'ko' | 'es'; label: string }[] = [
+    { code: 'ko', label: t('langKo') },
+    { code: 'es', label: t('langEs') },
+  ];
+
+  const langDropdownProps = {
+    language,
+    isLangOpen,
+    setIsLangOpen,
+    setLanguage,
+    selectLangLabel: t('selectLang'),
+    langs,
+  };
+
+  return (
+    <>
+      <DesktopHeader
+        searchValue={searchValue}
+        setSearchValue={setSearchValue}
+        handleSearch={handleSearch}
+        langDropdownProps={langDropdownProps}
+        handleLogout={handleLogout}
+        isHome={isHome}
+        isKo={isKo}
+        activeCategory={activeCategory}
+        handleCategoryClick={handleCategoryClick}
+      />
+
+      <MobileHeader
+        mobileHeaderRef={mobileHeaderRef}
+        isSearchOpen={isSearchOpen}
+        setIsSearchOpen={setIsSearchOpen}
+        searchValue={searchValue}
+        setSearchValue={setSearchValue}
+        handleSearch={handleSearch}
+        isKo={isKo}
+        langDropdownProps={langDropdownProps}
+        isHome={isHome}
+        activeCategory={activeCategory}
+        handleCategoryClick={handleCategoryClick}
+      />
+
+      <MobileTopBar isKo={isKo} />
+    </>
+  );
+}
