--- src/widgets/header/ui/HeaderTop.tsx
@@ -1,5 +1,0 @@
-const HeaderTop = () => {
-  return <div></div>;
-};
-
-export default HeaderTop;
